+++ cli/src/lib/compiler/entrypoints.js
@@ -1,6 +1,7 @@
 const path = require('path')
 const { reporter, chalk } = require('@dhis2/cli-helpers-engine')
 const fs = require('fs-extra')
+const { isApp } = require('../parseConfig')
 
 const verifyEntrypoint = ({ entrypoint, basePath, resolveModule }) => {
     if (!entrypoint.match(/^(\.\/)?src\//)) {
@@ -25,7 +26,7 @@ exports.verifyEntrypoints = ({
     paths,
     resolveModule = require.resolve,
 }) => {
-    if (config.type === 'app') {
+    if (isApp(config.type)) {
         if (
             !config.entryPoints ||
             (!config.entryPoints.app && !config.entryPoints.plugin)